+++ src/devices/gledopto.ts
@@ -417,8 +417,8 @@ const definitions: DefinitionWithExtend[] = [
         model: 'GL-C-003P',
         vendor: 'Gledopto',
         ota: ota.zigbeeOTA,
-        description: 'Zigbee LED Controller RGB (pro)',
-        extend: [light({color: {modes: ['xy', 'hs'], enhancedHue: true}}), identify(), gledoptoConfigureReadModelID()],
+        description: 'Zigbee LED Controller CCT (pro)',
+        extend: [light({colorTemp: {range: [158, 500]}}), identify(), gledoptoConfigureReadModelID()],
     },
     {
         zigbeeModel: ['GL-C-008P'],