+++ src/lib/philips.ts
@@ -1,6 +1,8 @@
 import {Zcl} from 'zigbee-herdsman';
 
+import fz from '../converters/fromZigbee';
 import tz from '../converters/toZigbee';
+import * as reporting from '../lib/reporting';
 import {ColorRGB, ColorXY} from './color';
 import * as libColor from './color';
 import * as exposes from './exposes';
@@ -8,9 +10,9 @@ import {logger} from './logger';
 import * as modernExtend from './modernExtend';
 import * as ota from './ota';
 import * as globalStore from './store';
-import {Fz, KeyValue, KeyValueAny, Tz} from './types';
+import {Configure, Fz, KeyValue, KeyValueAny, ModernExtend, Tz} from './types';
 import * as utils from './utils';
-import {isObject} from './utils';
+import {exposeEndpoints, isObject} from './utils';
 
 const NS = 'zhc:philips';
 const ea = exposes.access;
@@ -71,12 +73,15 @@ export function philipsLight(args?: modernExtend.LightArgs & {hueEffect?: boolea
             }
             result.exposes.push(
                 // gradient_scene is deprecated, use gradient instead
-                e.enum('gradient_scene', ea.SET, Object.keys(gradientScenes)),
-                e
-                    .list('gradient', ea.ALL, e.text('hex', ea.ALL).withDescription('Color in RGB HEX format (eg #663399)'))
-                    .withLengthMin(1)
-                    .withLengthMax(9)
-                    .withDescription('List of RGB HEX colors'),
+                ...exposeEndpoints(e.enum('gradient_scene', ea.SET, Object.keys(gradientScenes)), args.endpointNames),
+                ...exposeEndpoints(
+                    e
+                        .list('gradient', ea.ALL, e.text('hex', ea.ALL).withDescription('Color in RGB HEX format (eg #663399)'))
+                        .withLengthMin(1)
+                        .withLengthMax(9)
+                        .withDescription('List of RGB HEX colors'),
+                    args.endpointNames,
+                ),
             );
             result.configure.push(async (device, coordinatorEndpoint, definition) => {
                 for (const ep of device.endpoints.filter((ep) => ep.supportsInputCluster('manuSpecificPhilips2'))) {
@@ -85,7 +90,7 @@ export function philipsLight(args?: modernExtend.LightArgs & {hueEffect?: boolea
             });
         }
         effects.push('finish_effect', 'stop_effect', 'stop_hue_effect');
-        result.exposes.push(e.enum('effect', ea.SET, effects));
+        result.exposes.push(...exposeEndpoints(e.enum('effect', ea.SET, effects), args.endpointNames));
     }
     return result;
 }
@@ -94,6 +99,31 @@ export function philipsOnOff(args?: modernExtend.OnOffArgs) {
     args = {powerOnBehavior: false, ota: ota.zigbeeOTA, ...args};
     const result = modernExtend.onOff(args);
     result.toZigbee.push(philipsTz.hue_power_on_behavior, philipsTz.hue_power_on_error);
+    return result;
+}
+
+export function philipsTwilightOnOff() {
+    const fromZigbee = [fz.ignore_command_on, fz.ignore_command_off, fz.hue_twilight];
+    const exposes = [
+        e.action([
+            'dot_press',
+            'dot_hold',
+            'dot_press_release',
+            'dot_hold_release',
+            'hue_press',
+            'hue_hold',
+            'hue_press_release',
+            'hue_hold_release',
+        ]),
+    ];
+    const toZigbee: Tz.Converter[] = [];
+    const configure: Configure[] = [
+        async (device, coordinatorEndpoint) => {
+            const endpoint = device.getEndpoint(1);
+            await reporting.bind(endpoint, coordinatorEndpoint, ['genOnOff', 'manuSpecificPhilips']);
+        },
+    ];
+    const result: ModernExtend = {exposes, fromZigbee, toZigbee, configure, isModernExtend: true};
     return result;
 }